refactor(notifications): consolidate routes to NotificationCenterController
- Update routes/web.php to use NotificationCenterController for all notification actions
- Remove legacy app/Http/Controllers/NotificationController.php
